refactor: InMemoryUserStorage 패키지 위치를 domain으로 변경
