+++ src/main/java/codesquad/domain/InMemoryUserStorage.java
@@ -1,4 +1,4 @@
-package codesquad;
+package codesquad.domain;
 
 import codesquad.domain.model.User;
 import java.util.Map;
